android: Move CoreErrorDialogFragment to its own file
